Add files via upload
This was a demo project to create a simple tic-tac-toe game with AI abilities.  I started with the tic-tac-toe code base used in the React.js tutorial.  I then augmented the code with additional UX class components, and utilized JavaScript Promises to handle asynchronous waits.  I include the code for the Minimax move calculation algorithm, which required a bit of code modification and debugging.